chatapp/requirements.txt
streamlit==1.21.0
Pillow==9.2.0
langchain==0.0.176
python-dotenv==0.21.0
faiss-cpu==1.7.3